// Gestao_Hospitalar/src/app/Models/Produto.ts
export interface Produto {
    referencia: string; // Referência do produto
    descricao: string; // Descrição do produto
    familia: number; // Família do produto
    subfamilia: number; // Subfamília do produto
    unidadeEntrada: number; // Unidade de entrada
    unidadeSaida: number; // Unidade de saída
 
  }
  